Update Snake.js - h/t Nokia 2.10
As part of beginners coding challenge I changed the rules so that if the snake travels through the outbounds of the canvas it will emerge the other side and the game will continue. 
This emulates Nokia's second iteration of the game they loaded on to their hardware.
Also changing the background color in the index.html file to mark the versions out as different.